+++ token_types.py
@@ -5,7 +5,7 @@ from enum import Enum
 Enum containing all the token types.
 '''
 class Token_types(Enum):
-    INTEGER  = "INTEGER"
+    FLOAT    = "FLOAT"
     PLUS     = "HORIZONTAL_AND_VERTICAL_LINES"
     MINUS    = "HORIZONTAL_LINE"
     DIVIDE   = "DIAGONAL_LINE"
